♻️ #5 Update the directory structure
不要になったコードを削除し、ディレクトリ構造を整理する。
