+++ packages/common/action-builder/src/index.ts
@@ -1,3 +1,2 @@
-export * from "./core";
 export * from "./generate";
 export * from "./define-action";
